# [WIP] ICP: AES_GCM: add support for SSE4.1 acceleration