--- a/AUTHORS
+++ b/AUTHORS
@@ -49,6 +49,7 @@ CONTRIBUTORS:
     Arkadiusz Bubała <arkadiusz.bubala@open-e.com>
     Arne Jansen <arne@die-jansens.de>
     Aron Xu <happyaron.xu@gmail.com>
+    Attila Fülöp <attila@fueloep.org>
     Bart Coddens <bart.coddens@gmail.com>
     Basil Crow <basil.crow@delphix.com>
     Huang Liu <liu.huang@zte.com.cn>