… Makefile.am
… contrib/icp/gcm-simd/isa-l_crypto-ported/gcm_sse.S
… contrib/icp/gcm-simd/isa-l_crypto/gcm128_sse.asm
… contrib/icp/gcm-simd/isa-l_crypto/gcm256_sse.asm
… contrib/icp/gcm-simd/isa-l_crypto/gcm_defines.asm
… contrib/icp/gcm-simd/isa-l_crypto/gcm_sse.asm
… contrib/icp/gcm-simd/isa-l_crypto/reg_sizes.asm
… lib/libicp/Makefile.am
… module/Kbuild.in
… module/icp/algs/modes/gcm.c
… module/icp/algs/modes/modes.c
… module/icp/asm-x86_64/modes/THIRDPARTYLICENSE.intel
… module/icp/asm-x86_64/modes/THIRDPARTYLICENSE.intel.descrip
… module/icp/asm-x86_64/modes/isalc_gcm128_sse.S
… module/icp/asm-x86_64/modes/isalc_gcm256_sse.S
… module/icp/asm-x86_64/modes/isalc_gcm_defines.S
… module/icp/asm-x86_64/modes/isalc_gcm_sse.S
… module/icp/asm-x86_64/modes/isalc_reg_sizes.S
… module/icp/include/modes/modes.h
… module/icp/io/aes.c