### Motivation and Context

Small NAS solution are often based on hardware with no AVX support, where encryption performs subpar. Most of them do support  SSE4.1 on the other hand. Supporting SSE4.1  will give them a massive boost in aes-gcm crypto performance.


### Description

This is still a draft, a detailed description will be added once this is ready for merging.

Closes #10846

### How Has This Been Tested?

Inside a VM two `fio` invocations on an encrypted datasets and `mprime -t` were run in parallel. Various other tests
were done as well. Finally, this PR contains debugging code which runs  the AVX and SSE implementations on the same context an compares the results (See the DEBUG_GCM_ASM  #define).

### Types of changes
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [x] Performance enhancement (non-breaking change which improves efficiency)
- [ ] Code cleanup (non-breaking change which makes code smaller or more readable)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Library ABI change (libzfs, libzfs\_core, libnvpair, libuutil and libzfsbootenv)
- [ ] Documentation (a change to man pages or other documentation)

### Checklist:
- [x] My code follows the OpenZFS [code style requirements](https://github.com/openzfs/zfs/blob/master/.github/CONTRIBUTING.md#coding-conventions).
- [ ] I have updated the documentation accordingly.
- [x] I have read the [**contributing** document](https://github.com/openzfs/zfs/blob/master/.github/CONTRIBUTING.md).
- [ ] I have added [tests](https://github.com/openzfs/zfs/tree/master/tests) to cover my changes.
- [ ] I have run the ZFS Test Suite with this change applied.
- [x] All commit messages are properly formatted and contain [`Signed-off-by`](https://github.com/openzfs/zfs/blob/master/.github/CONTRIBUTING.md#signed-off-by).