@AttilaFueloep 2023-02-25
The people at ixsystems may be interested in testing this too. Surely they'll have a far better test infrastructure to test this on than I do.