@AttilaFueloep 2023-02-25
One important point regarding testing I missed above:

If you are testing this on hardware which has AVX support you need to activate SSE support by doing

```sh
# echo  sse4_1 >/sys/module/zfs/parameters/icp_gcm_impl
```
or setting it on module load time.