@lundman 2023-02-25
Excellent, I will give it a go on macOS and Windows.