@robn 2023-02-25
I did a light fio run on a small Celeron J3455 [Goldmont] board with SSE4.2 but no AVX. Using a single-vdev tmpfs-backed pool, no compression:

* `generic`: ~3.7MB/s
* `pclmulqdq`: ~6.1MB/s
* `sse4_1`: ~13.2MB/s
* no encryption: ~17MB/s

fio was: `fio --name=xxx --size=4K --bs=4K --rw=randrw --ioengine=psync --sync=1 --iodepth=32 --numjobs=1 --direct=
1 --end_fsync=1 --gtod_reduce=1 --time_based --runtime=60`

Hardly science, but those initial numbers seem pretty good. Strong work!