@AttilaFueloep 2023-02-26
@oberien Well, I'd suggest the following:

Spin up a VM which has all the dependencies for building ZFS installed, but not the ZFS packages itself.

Inside the VM do:
```sh
git clone http://github.com/openzfs/zfs/
cd zfs
git pull origin pull/14531/head
./autogen.sh && ./configure --prefix=/ --enable-debuginfo --enable-debug && make 
```
Open a new root terminal inside the VM since for me the installation step  failed with sudo. 

```sh
cd <dir where you build zfs>
make install
modprobe zfs
echo sse4_1 >/sys/module/zfs/parameters/icp_gcm_impl
zpool create ...
zfs create -o encryption= ...
```
Run your favorite workloads and benchmark tools.